requirements.txt
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
linearmodels==6.0
statsmodels==0.14.2
matplotlib==3.9.0
python-pptx==0.6.23
requests==2.32.3
pyarrow==16.1.0
openpyxl==3.1.4
scikit-learn==1.5.0
seaborn==0.13.2
tqdm==4.66.4
